backend/requirements.txt
flask
flask-cors
pandas
PyPDF2
faiss-cpu
sentence-transformers
requests
